+++ source/Nevermore/Advanced/ThreadSafeAsyncEnumerable.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
+using Nevermore.Advanced.Concurrency;
 
 namespace Nevermore.Advanced
 {
