+++ source/Nevermore/Advanced/ThreadSafeEnumerable.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using Nevermore.Advanced.Concurrency;
 
 namespace Nevermore.Advanced
 {
